Separate downloadclient & httpclient
